+++ docs/app/Examples/collections/Breadcrumb/Variations/BreadcrumbBigSizeExample.js
@@ -1,0 +1,14 @@
+import React from 'react'
+import { Breadcrumb } from 'stardust'
+
+const BreadcrumbBigSizeExample = () => (
+  <Breadcrumb size='big'>
+    <Breadcrumb.Section link>Home</Breadcrumb.Section>
+    <Breadcrumb.Divider icon='right chevron' />
+    <Breadcrumb.Section link>Registration</Breadcrumb.Section>
+    <Breadcrumb.Divider icon='right chevron' />
+    <Breadcrumb.Section active>Personal Information</Breadcrumb.Section>
+  </Breadcrumb>
+)
+
+export default BreadcrumbBigSizeExample