+++ docs/app/Examples/modules/Dropdown/index.js
@@ -1,0 +1,14 @@
+import React, { Component } from 'react'
+import Content from './Content'
+import States from './States'
+
+export default class DropdownExamples extends Component {
+  render() {
+    return (
+      <div>
+        <Content />
+        <States />
+      </div>
+    )
+  }
+}