+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Change Log
 
+## [Unreleased](https://github.com/TechnologyAdvice/stardust/tree/HEAD)
+
+[Full Changelog](https://github.com/TechnologyAdvice/stardust/compare/v0.24.1...HEAD)
+
+**Merged pull requests:**
+
+- Add doc site routes [\#337](https://github.com/TechnologyAdvice/stardust/pull/337) ([levithomason](https://github.com/levithomason))
+
+## [v0.24.1](https://github.com/TechnologyAdvice/stardust/tree/v0.24.1) (2016-07-17)
+[Full Changelog](https://github.com/TechnologyAdvice/stardust/compare/v0.24.0...v0.24.1)
+
+## [v0.24.0](https://github.com/TechnologyAdvice/stardust/tree/v0.24.0) (2016-07-17)
+[Full Changelog](https://github.com/TechnologyAdvice/stardust/compare/v0.23.0...v0.24.0)
+
+**Implemented enhancements:**
+
+- \<Breadcrumb /\> Component [\#184](https://github.com/TechnologyAdvice/stardust/issues/184)
+- Add Breadcrumb component [\#321](https://github.com/TechnologyAdvice/stardust/pull/321) ([layershifter](https://github.com/layershifter))
+
 ## [v0.23.0](https://github.com/TechnologyAdvice/stardust/tree/v0.23.0) (2016-07-16)
 [Full Changelog](https://github.com/TechnologyAdvice/stardust/compare/v0.22.0...v0.23.0)